rapido/requirements.txt
aiohttp==3.8.6
websockets==11.0.3
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.25.2
python-dotenv==1.0.0
elevenlabs==0.2.26
json5==0.9.14
ffmpeg-python==0.2.0
librosa==0.10.1
soundfile==0.12.1
requests==2.31.0
httpx==0.24.1
